+++ app/server/handlers_test.go
@@ -5,7 +5,6 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"strings"
-	"testing"
 
 	"github.com/gorilla/mux"
 )
@@ -39,28 +38,4 @@ func genJson(str interface{}) string {
 	body, _ := json.Marshal(str)
 
 	return string(body)
-}
-
-func TestSimpleCreate(t *testing.T) {
-	type input struct {
-		Name string `json:"name"`
-	}
-
-	name := "test"
-
-	inputBody := &input{
-		Name: name,
-	}
-
-	body := genJson(inputBody)
-
-	res := reqRes(genSrv(), "POST", "/simple/create", body)
-
-	if code := res.Code; code != 200 {
-		t.Errorf("response code = %v != 200", code)
-	}
-
-	if body := res.Body.String(); body != name {
-		t.Errorf("expected %v, got %v", name, body)
-	}
 }
